+++ Classes/PBUnsortableTableHeader.h
@@ -12,7 +12,7 @@
 @interface PBUnsortableTableHeader : NSTableHeaderView {
 	IBOutlet NSArrayController *controller;
 	int clickCount;
-	int columnIndex;
+	NSInteger columnIndex;
 }
 
 @end
